with 03_ex

+++ 02_exercise/mass_of_blocks.cpp
@@ -6,7 +6,7 @@ using namespace std;
 int main() {
     long long n, density;
     cin >> n >> density;
-    vector<unsigned int> sizes;
+    vector<unsigned long long> sizes;
     for (int i = 0; i < n; ++i) {
         unsigned long long mul = 1;
         for (int j = 0; j < 3; ++j) {
